+++ client/coral-admin/src/routes/Community/components/Community.js
@@ -3,33 +3,22 @@ import PropTypes from 'prop-types';
 import styles from './Community.css';
 import People from '../containers/People';
 import CommunityMenu from '../containers/CommunityMenu';
-import RejectUsernameDialog from './RejectUsernameDialog';
+import RejectUsernameDialog from '../containers/RejectUsernameDialog';
 import FlaggedAccounts from '../containers/FlaggedAccounts';
 
 class Community extends Component {
   renderTab() {
-    const { route, community, ...props } = this.props;
+    const { route } = this.props;
     const activeTab = route.path === ':id' ? 'flagged' : route.path;
 
     if (activeTab === 'people') {
-      return (
-        <People
-          community={community}
-          data={this.props.data}
-          root={this.props.root}
-        />
-      );
+      return <People />;
     }
 
     return (
       <div>
-        <FlaggedAccounts data={this.props.data} root={this.props.root} />
-        <RejectUsernameDialog
-          user={community.user}
-          open={community.rejectUsernameDialog}
-          rejectUsername={props.rejectUsername}
-          handleClose={props.hideRejectUsernameDialog}
-        />
+        <FlaggedAccounts />
+        <RejectUsernameDialog />
       </div>
     );
   }
@@ -46,11 +35,6 @@ class Community extends Component {
 
 Community.propTypes = {
   route: PropTypes.object,
-  community: PropTypes.object,
-  rejectUsername: PropTypes.func.isRequired,
-  hideRejectUsernameDialog: PropTypes.func.isRequired,
-  data: PropTypes.object,
-  root: PropTypes.object,
 };
 
 export default Community;
