+++ client/coral-admin/src/routes/Community/index.js
@@ -1,0 +1,1 @@
+export { default } from './components/Community';
